Move readerRevenueApis under newPaymentFlow
